Make all mounts to be performed in a private namespace.
This makes mount and umount commands unnecessary because all mounts
are done in a private namespace during enter or exec phases.
Also, enter and exec commands are unnecessary as well because
action to be performed can be determined by presence of command name and args.
Hence, no action is needed to be specified;
the new command line syntax looks like this:
mer-sdk-chroot [command args...]

Signed-off-by: Oleg Girko <ol+mer@infoserver.ru>

Added some compatibility code so sdk mount/umount is an error and
sdk enter/exec still works

Modified-by: David Greaves <david@dgreaves.com>