mb: Add multiple job support.
A lot of packaging uses %jobs as an argument to make (thanks to spectacle), but
seemingly this isn't automagically set up, so we need to do it ourselves.

This makes builds actually use many cores, and go a lot faster as a result.

This can be specified by the user (-j <jobs>), or else the number of processors
is used as a default.