[sdk-manage] Add support for ssu register
To register the build engine and all sb2 targets that require
registration:
$ sdk-manage --register-all --user <username> --password <password>

To register one, multiple or all sb2 targets:
$ sdk-manage --target --register --user <uname> --password <pass> --all | <tname> [<tname>...]

To register the build engine:
$ sdk-manage --sdk --register --user <name> --password <pass>

In addition to the register changes, adjust listing targets so that
only targets that are accessible from where sdk-manage is called are
listed.

The accessibility check is required because the user can have sb2
targets that are created on filesystems not accessible by the build
engine.