[sdk-manage] Fix targets dir not writable for sdk user. Contribute to JB#38881
The original command would fail, leaving the task to create targets dir
for the `sudo mkdir -p` command below in the function, which would
create the directory owned by root. All this can happen unnoticed only
when installing target from a local URL so that it does not need to
store the local copy of the tarball inside the targets dir before it is
actually created - which is the case when using setup-buildengine.sh
from sdk-build-tools.