[mb2] Improve --fix-version. Contribute to JB#38957
Make it behave more like tar_git does for commits ahead of the last tag.
tar_git relies on "token" coming from webhook configuration - do not
require this and allow to directly select tag on command line as a
fallback. Also handle the case when mb2 is used with a working tree in
dirty state - create a stash in this case.