[mb2] Make use of rpmbuild's --short-circuit option. Contribute to JB#39755
The --do* options are Mer extensions to rpmbuild. Nowadays (nearly) the
same effect can be achieved with standard rpmbuild options.

The --short-circuit option when used with -bb (which is true for `mb2
rpm`) has the effect of marking packages by adding an unsatisfiable
dependency 'rpmlib(ShortCircuited)'. This is to prevent accidental use
of short-circuited RPMs.  A developer using this feature for local
development and testing is supposed to either use "--nodeps" or create
an empty package to satisfy this dependency.