[sdk-setup-enginelan] Prevent ConnMan setting an invalid gateway.  Contribute to JB#41668
Using the SDK on a machine without internet connection available, the
original configuration led to default gateway being set for this
interface, which is the interface connected to the virtual network for
communication between build engine and emulator. I haven't found a
better way to fix this. Tried reordering the services, using
AlwaysConnectedTechnologies and EnableOnlineCheck config options and
more, with no effect.