mb2: Add option to not update target dependencies.
Adds option -n or --no-deps

Sometimes it is useful to do the mb2 operations on a target without
updating package dependencies (for example, having installed
locally built version of dependency for testing). Add option to
disable the dependency updating so that target stays in proper state
for the duration of build etc.

For example to build without updating dependencies
mb2 -t target -n build

[mb2] Add option to not update target dependencies. Fixes JB#42728