[mb2] Pass source path as the last arg to cmake. JB#47262
CMake does very poor job at argument parsing. Some options, including
"-C", are processed such a way that their arguments overwrite
path-to-source argument if it occurred earlier on the command line.

E.g. when invoked like in

  cmake ../source -C my-cache.cmake

cmake would fail complaining that "The source directory 'my-cache.cmake'
is a file, not a directory.

This is clearly a bug in CMake. Even if fixed, we need to live with that
in order to support older build targets.