[mb2] Fix determining latest rpms
It seems zypper does not really sorts output by version. This was
detected by the self-test when executed inside build engine. It seems
the ordering is influenced by the order of the respective directory
entries under the repo directory. Under build engine this is on tmpfs
and this seems to make the difference.