[packaging] Fix calling systemctl from %post. JB#50670
systemctl is located under /usr/bin now. There has been a global
systemd_post macro available that invokes systemctl without path. The
only difference is that it actually does restrict execution to the
initial installation (compared to the code here which only claims that
in the comment). We do not normally upgrade build engine, so this
difference should not cause any trouble.