[sdk-manage] Use less for help and refactor usage output. Fixes JB#52140
Initial motivation was to just change long help output from stderr to
stdout so that it is easier to grep the output but as other utils seem
to utilize less for long help change to that instead. This way it's
still easy to grep the help if needed but we use less by default.

Signed-off-by: Juho Hämäläinen <juho.hamalainen@jolla.com>