[git-change-log] Only consider tags reachable through first parent
The original implementation relied on the assumption that tags on all
branches strictly follow the same naming conventions. That assumption
easily breaks in the real world and even in that ideal situation it
didn't yield ideal results (see the updated test case). Dealing with
relevant tags appearing on side branches is difficult to do right, so we
should have a use case before trying to support that.